requirements.txt
django>=6.0
djangorestframework>=3.17
pillow>=12.0
django-cors-headers>=4.9
django-filter>=25.0
stripe>=15.0
celery>=5.6
redis>=7.4
django-storages>=1.14
boto3>=1.42
psycopg2-binary>=2.9
django-allauth>=65.0
django-crispy-forms>=2.6
crispy-bootstrap5>=2026.3
django-mptt>=0.18
django-taggit>=6.0
django-money>=3.6
whitenoise>=6.12
dj-database-url>=3.1
python-decouple>=3.8
django-import-export>=4.4
reportlab>=4.4
qrcode>=8.0
